backend/requirements.txt
flask
flask-cors
onnxruntime
opencv-python
Pillow
google-genai
python-dotenv
numpy
werkzeug
